fix(agent-cli): split the catalogue generator's cli entry point out
gen:agent-catalogue/gen:agent-catalogue:check ran gen-agent-catalogue.ts
directly, guarded by an `if (resolve(process.argv[1]) ===
fileURLToPath(import.meta.url))` self-invocation check. If that
comparison ever stopped matching (a symlinked/shimmed invocation, or a
Windows drive-letter/short-path mismatch) the check would exit 0 having
done nothing, passing CI against a stale catalogue -- exactly the class
of bug gen-api-docs.cli.mjs was already split out to close, and the two
generators disagreed on how to solve the identical problem.

main is now exported with no self-invocation, and a separate
gen-agent-catalogue.cli.ts wrapper (mirroring gen-api-docs.cli.mjs) is
the only thing package.json's two scripts run. Verified live: pnpm
gen:agent-catalogue:check passes against the current tree and fails
against a planted drift, through the new wrapper.

Refs #1160
