+++ packages/shared/package.json
@@ -46,8 +46,8 @@
     "test:watch": "vitest",
     "gen:ipc": "tsx scripts/gen-ipc-rust.ts",
     "gen:ipc:check": "tsx scripts/gen-ipc-rust.ts --check",
-    "gen:agent-catalogue": "tsx scripts/gen-agent-catalogue.ts",
-    "gen:agent-catalogue:check": "tsx scripts/gen-agent-catalogue.ts --check"
+    "gen:agent-catalogue": "tsx scripts/gen-agent-catalogue.cli.ts",
+    "gen:agent-catalogue:check": "tsx scripts/gen-agent-catalogue.cli.ts --check"
   },
   "dependencies": {
     "franc": "^6.2.0",
